oscal-common-cp/pyproject.toml
[tool.poetry]
name = "oscal-common-cp"
version = "0.1.0"
description = ""
authors = ["Robert Sherwood <robert.sherwood@credentive.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
oscal-pydantic-v2 = "^2023.12.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
